+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to `homebridge` will be documented in this file. This projec
 - fix(matter): add missing switch cluster to generic switch
 - fix(matter): add missing smoke/co cluster to smoke sensor
 - fix(matter): add missing pump control cluster to pump
+- fix(matter): add missing thermostat cluster to room ac
 
 ### Homebridge Dependencies